functions/requirements.txt
google-cloud-firestore
requests